examples/pipelines/demo-document-indexing/requirements.txt
pathway==0.8.2
python-dotenv==1.0.1
litellm==1.17.3
unstructured[all-docs]==0.10.28
mpmath==1.3.0
